se_services: Improve semaphore usage
In the case of a semaphore timeout, the code attempts to reset
the semaphore to avoid a late incoming SE reply affecting the
next SE request(which will not block for send/receive completion
as the semaphores are already available from the late/timed out
replies for the earlier request). But there is still a large window
of time in which a late reply can come for a timed out SE service
request. Improve the usage by resetting the semaphore
just before sending the next request.

Signed-off-by: Silesh C V <silesh@alifsemi.com>